siirretty Datatype uuteen tiedostoon
