+++ src/org/kaivos/röda/commands/StreamPopulator.java
@@ -5,9 +5,9 @@ import java.util.Collections;
 import java.util.function.Supplier;
 
 import org.kaivos.röda.Builtins;
+import org.kaivos.röda.Datatype;
 import org.kaivos.röda.Interpreter;
 import org.kaivos.röda.Interpreter.RödaScope;
-import org.kaivos.röda.Parser.Datatype;
 import org.kaivos.röda.Parser.Parameter;
 import org.kaivos.röda.Parser.Record;
 import org.kaivos.röda.RödaStream;
